+++ eth/types/receipt.go
@@ -22,9 +22,9 @@ import (
 	"io"
 	"unsafe"
 
-	"github.com/HPISTechnologies/3rd-party/eth/common"
-	"github.com/HPISTechnologies/3rd-party/eth/common/hexutil"
-	"github.com/HPISTechnologies/3rd-party/eth/rlp"
+	"github.com/arcology/3rd-party/eth/common"
+	"github.com/arcology/3rd-party/eth/common/hexutil"
+	"github.com/arcology/3rd-party/eth/rlp"
 )
 
 //go:generate gencodec -type Receipt -field-override receiptMarshaling -out gen_receipt_json.go
@@ -55,6 +55,7 @@ type Receipt struct {
 	TxHash          common.Hash    `json:"transactionHash" gencodec:"required"`
 	ContractAddress common.Address `json:"contractAddress"`
 	GasUsed         uint64         `json:"gasUsed" gencodec:"required"`
+	SpawnedTxHash   common.Hash    `json:"spawned transactionHash"`
 }
 
 type receiptMarshaling struct {
@@ -74,6 +75,7 @@ type receiptRLP struct {
 	TxHash          common.Hash
 	ContractAddress common.Address
 	GasUsed         uint64
+	SpawnedTxHash   common.Hash
 }
 
 type receiptStorageRLP struct {
@@ -84,6 +86,7 @@ type receiptStorageRLP struct {
 	ContractAddress   common.Address
 	Logs              []*LogForStorage
 	GasUsed           uint64
+	SpawnedTxHash     common.Hash
 }
 
 // NewReceipt creates a barebone transaction receipt, copying the init fields.
@@ -100,7 +103,7 @@ func NewReceipt(root []byte, failed bool, cumulativeGasUsed uint64) *Receipt {
 // EncodeRLP implements rlp.Encoder, and flattens the consensus fields of a receipt
 // into an RLP stream. If no post state is present, byzantium fork is assumed.
 func (r *Receipt) EncodeRLP(w io.Writer) error {
-	return rlp.Encode(w, &receiptRLP{r.statusEncoding(), r.CumulativeGasUsed, r.Bloom, r.Logs, r.TxHash, r.ContractAddress, r.GasUsed})
+	return rlp.Encode(w, &receiptRLP{r.statusEncoding(), r.CumulativeGasUsed, r.Bloom, r.Logs, r.TxHash, r.ContractAddress, r.GasUsed, r.SpawnedTxHash})
 }
 
 // DecodeRLP implements rlp.Decoder, and loads the consensus fields of a receipt
@@ -113,7 +116,7 @@ func (r *Receipt) DecodeRLP(s *rlp.Stream) error {
 	if err := r.setStatus(dec.PostStateOrStatus); err != nil {
 		return err
 	}
-	r.CumulativeGasUsed, r.Bloom, r.Logs, r.TxHash, r.ContractAddress, r.GasUsed = dec.CumulativeGasUsed, dec.Bloom, dec.Logs, dec.TxHash, dec.ContractAddress, dec.GasUsed
+	r.CumulativeGasUsed, r.Bloom, r.Logs, r.TxHash, r.ContractAddress, r.GasUsed, r.SpawnedTxHash = dec.CumulativeGasUsed, dec.Bloom, dec.Logs, dec.TxHash, dec.ContractAddress, dec.GasUsed, dec.SpawnedTxHash
 	return nil
 }
 
@@ -168,6 +171,7 @@ func (r *ReceiptForStorage) EncodeRLP(w io.Writer) error {
 		ContractAddress:   r.ContractAddress,
 		Logs:              make([]*LogForStorage, len(r.Logs)),
 		GasUsed:           r.GasUsed,
+		SpawnedTxHash:     r.SpawnedTxHash,
 	}
 	for i, log := range r.Logs {
 		enc.Logs[i] = (*LogForStorage)(log)
@@ -192,7 +196,7 @@ func (r *ReceiptForStorage) DecodeRLP(s *rlp.Stream) error {
 		r.Logs[i] = (*Log)(log)
 	}
 	// Assign the implementation fields
-	r.TxHash, r.ContractAddress, r.GasUsed = dec.TxHash, dec.ContractAddress, dec.GasUsed
+	r.TxHash, r.ContractAddress, r.GasUsed, r.SpawnedTxHash = dec.TxHash, dec.ContractAddress, dec.GasUsed, dec.SpawnedTxHash
 	return nil
 }